Restart Waybar using its systemd service if it's enabled
For uwsm users, Hyprland recommends enabling Waybar via its Systemd service. The script will now check if the user is running Waybar using the Systemd service or not, if the service is not enabled, the script will restart Waybar like usual.

+++ install.sh
@@ -47,9 +47,14 @@ main() {
 
 	cprintf blue "\nRestarting Waybar..."
 
-	pkill waybar
-	waybar &> /dev/null &
-	disown
+	if systemctl --user list-unit-files waybar.service &>/dev/null &&
+		systemctl --user is-enabled waybar.service &>/dev/null; then
+		systemctl --user restart waybar.service
+	else
+		pkill waybar
+		waybar &> /dev/null &
+		disown
+	fi
 
 	if ((ERRORS > 0)); then
 		cprintf red "\nInstallation completed with $ERRORS errors"
